lesson-related models moved to lesson directory, location-related models moved to location directory
