+++ app/src/main/java/com/nailorsh/repeton/common/data/models/Id.kt
@@ -1,0 +1,4 @@
+package com.nailorsh.repeton.common.data.models
+
+@JvmInline
+value class Id(val value: String)
