+++ app/src/main/java/com/nailorsh/repeton/common/data/models/lesson/Lesson.kt
@@ -1,4 +1,4 @@
-package com.nailorsh.repeton.common.data.models
+package com.nailorsh.repeton.common.data.models.lesson
 
 import java.time.LocalDateTime
