+++ app/src/main/java/com/nailorsh/repeton/common/data/models/lesson/Subject.kt
@@ -1,4 +1,4 @@
-package com.nailorsh.repeton.common.data.models
+package com.nailorsh.repeton.common.data.models.lesson
 
 data class Subject(
     val id : String,
