+++ app/src/main/java/com/nailorsh/repeton/common/data/models/location/City.kt
@@ -1,0 +1,8 @@
+package com.nailorsh.repeton.common.data.models.location
+
+import com.nailorsh.repeton.common.data.models.Id
+
+data class City(
+    val id : Id,
+    val name : String,
+)
