+++ app/src/main/java/com/nailorsh/repeton/common/data/models/location/Location.kt
@@ -1,4 +1,4 @@
-package com.nailorsh.repeton.common.data.models.user
+package com.nailorsh.repeton.common.data.models.location
 
 data class Location(
     val id : String,
