+++ app/src/main/java/com/nailorsh/repeton/common/data/models/user/UserOptions.kt
@@ -1,0 +1,2 @@
+package com.nailorsh.repeton.common.data.models.user
+
