+++ app/src/main/java/com/nailorsh/repeton/features/currentlesson/data/CurrentLessonRepository.kt
@@ -1,6 +1,6 @@
 package com.nailorsh.repeton.features.currentlesson.data
 
-import com.nailorsh.repeton.common.data.models.Lesson
+import com.nailorsh.repeton.common.data.models.lesson.Lesson
 
 interface CurrentLessonRepository {
     suspend fun getLesson(id: Int): Lesson
