+++ app/src/main/java/com/nailorsh/repeton/features/currentlesson/presentation/ui/components/HomeworkCard.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import com.nailorsh.repeton.R
-import com.nailorsh.repeton.common.data.models.Homework
+import com.nailorsh.repeton.common.data.models.lesson.Homework
 
 
 @Composable
