+++ app/src/main/java/com/nailorsh/repeton/features/newlesson/data/FakeNewLessonRepository.kt
@@ -1,8 +1,8 @@
 package com.nailorsh.repeton.features.newlesson.data
 
 
-import com.nailorsh.repeton.common.data.models.Lesson
-import com.nailorsh.repeton.common.data.models.Subject
+import com.nailorsh.repeton.common.data.models.lesson.Lesson
+import com.nailorsh.repeton.common.data.models.lesson.Subject
 import com.nailorsh.repeton.common.data.sources.FakeLessonSource
 import com.nailorsh.repeton.common.data.sources.FakeSubjectsSource
 import kotlinx.coroutines.Dispatchers
