+++ app/src/main/java/com/nailorsh/repeton/features/newlesson/data/NewLessonRepository.kt
@@ -1,7 +1,7 @@
 package com.nailorsh.repeton.features.newlesson.data
 
-import com.nailorsh.repeton.common.data.models.Lesson
-import com.nailorsh.repeton.common.data.models.Subject
+import com.nailorsh.repeton.common.data.models.lesson.Lesson
+import com.nailorsh.repeton.common.data.models.lesson.Subject
 
 interface NewLessonRepository {
