+++ app/src/main/java/com/nailorsh/repeton/features/schedule/data/ScheduleRepository.kt
@@ -1,6 +1,6 @@
 package com.nailorsh.repeton.features.schedule.data
 
-import com.nailorsh.repeton.common.data.models.Lesson
+import com.nailorsh.repeton.common.data.models.lesson.Lesson
 import java.time.LocalDate
 
 interface ScheduleRepository {
