+++ app/src/main/java/com/nailorsh/repeton/features/schedule/presentation/ui/components/LessonCard.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.nailorsh.repeton.R
-import com.nailorsh.repeton.common.data.models.Lesson
+import com.nailorsh.repeton.common.data.models.lesson.Lesson
 import java.time.format.DateTimeFormatter
 import java.time.format.FormatStyle
